+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - added notes to daily graphics
 - added pdf for profiles
+- added parameter for reverse sortorder in daily graphics
  
 ### Changed
 - made request for valid url more flexible